🔧 CRUD operations ended (KVDB)

+++ entities/OpenHouseEventDynamic/db_open_house_event_dynamic.py
@@ -9,6 +9,8 @@ class OpenHouseEventDynamicDB:
         self.open_house_event_dynamic = open_house_event_dynamic
     
     def get_open_house_event_dynamic_by_property(self, property_id:int) -> OpenHouseEventDynamic:
+        if not property_id:
+            return None
         redis_client = get_redis_client()
         raw_data = redis_client.get(f"dynamic_info:{property_id}")
         if not raw_data:
@@ -22,12 +24,40 @@ class OpenHouseEventDynamicDB:
         return OpenHouseEventDynamic(property_id=property_id, dynamic_info=dynamic_info)
     
     def delete_open_house_event_dynamic_by_property(self, property_id: int) -> bool:
+        if not property_id:
+            return False
         redis_client = get_redis_client()
         result = redis_client.delete(f"dynamic_info:{property_id}")
         return bool(result)
-        
-    # todo: implement the following methods
-    # create open house event (new property inserted in mongo)
-    # update open house event (open house date and time updated in mongo) -> mabye notify the buyers about the change
-    # delete open house event (property deleted in mongo) -> also delete all the reservations related to this property
-    # update open house event when the event occurs -> also delete all the reservations related to this property for this event
+    
+    def create_open_house_event_dynamic(self, property_id:int, dynamic_info:DynamicInfo) -> bool:
+        if not property_id:
+            return False
+        if not dynamic_info:
+            return False
+        redis_client = get_redis_client()
+        data = {
+            "event_id": dynamic_info.event_id,
+            "date": dynamic_info.date,
+            "time": dynamic_info.time
+        }
+        result = redis_client.set(f"dynamic_info:{property_id}", json.dumps(data))
+        return bool(result)
+    
+    def update_open_house_event_dynamic(self, property_id:int = None, dynamic_info:DynamicInfo = None) -> bool:
+        redis_client = get_redis_client()
+        raw_data = redis_client.get(f"dynamic_info:{property_id}")
+        if not raw_data:
+            return False
+        data = json.loads(raw_data)
+        if dynamic_info.event_id:
+            data["event_id"] = dynamic_info.event_id
+        if dynamic_info.date:
+            data["date"] = dynamic_info.date
+        if dynamic_info.time:
+            data["time"] = dynamic_info.time
+        result = redis_client.set(f"dynamic_info:{property_id}", json.dumps(data))
+        return bool(result)
+    
+    # when the ttl of the key expires, the event is considered closed and another dynamic event is created, updating the date of the previous event
+    # oss: open house event static is needed for the operation (time_schedule is needed)
